Initial commit
implemented clean arch

updated readme

added missing end of file newlines

Resolved Mati's comments
