--- lib/repository/interface/genre_interface.dart
@@ -1,6 +1,0 @@
-import '../../model/genre_data.dart';
-
-abstract class GenreInterface {
-  Future<List<Genre>> getGenres();
-  Future<List<String>> getGenresById(List<int> genreList);
-}
