--- lib/repository/interface/movie_interface.dart
@@ -1,6 +1,0 @@
-import '../../model/movie_data.dart';
-
-abstract class MovieInterface {
-  Future<Movie> getMovieById(int id);
-  Future<List<Movie>> getMovies();
-}
