+++ lib/src/core/utils/dimens.dart
@@ -1,0 +1,42 @@
+class Dimens {
+  // Movie List
+  static const posterHeight = 500.0;
+  static const posterPadding = 25.0;
+  static const posterRadius = 25.0;
+  static const cardMargin = 30.0;
+  static const cardRadiusTop = 25.0;
+  static const cardRadiusBottom = 10.0;
+
+  // Movie Detail
+  static const posterDetailYAligment = 20.0;
+  static const posterDetailXAligment = 10.0;
+  static const posterDetailHeight = 200.0;
+  static const posterDetailWidth = 140.0;
+  static const posterDetailStroke = 3.0;
+  static const posterDetailRadius = 5.0;
+  static const originalTitleSize = 22.0;
+  static const likesSpacing = 30.0;
+  static const releaseDateSize = 22.0;
+  static const releaseDatePadding = 20.0;
+  static const previewRadiusSize = 20.0;
+  static const footerSeparator = 20.0;
+  static const genresSeparator = 50.0;
+  static const genresCardRadius = 14.0;
+  static const genresTextSize = 16.0;
+
+  // Stars Constants
+  static const divideByHalf = 2;
+  static const maxRating = 5;
+  static const ratingTextSize = 25.0;
+
+  // Movie Information
+  static const defaultLikes = 1359;
+
+  // Size Constants
+  static const defaultTextSize = 18.0;
+  static const titleTextSize = 26.0;
+  static const infoPadding = 9.0;
+  static const titleStroke = 5.0;
+  static const titleStrokeMaxLines = 3;
+  static const zero = 0.0;
+}
