+++ lib/src/data/model/genre_model.dart
@@ -1,0 +1,15 @@
+import '../../domain/entity/genre_entity.dart';
+
+class GenreModel extends Genre {
+  GenreModel({
+    required super.id,
+    required super.name,
+  });
+
+  factory GenreModel.fromJson(Map<String, dynamic> genresJson) {
+    return GenreModel(
+      id: genresJson['id'] as int,
+      name: genresJson['name'] as String,
+    );
+  }
+}
