+++ lib/src/domain/entity/genre_entity.dart
@@ -1,0 +1,9 @@
+class Genre {
+  final int id;
+  final String name;
+
+  Genre({
+    required this.id,
+    required this.name,
+  });
+}
