+++ lib/src/domain/repository/interface/genre_interface.dart
@@ -1,0 +1,6 @@
+import '../../entity/genre_entity.dart';
+
+abstract class GenreInterface {
+  Future<List<Genre>> getGenres();
+  Future<List<String>> getGenresById(List<int> genreList);
+}
