+++ lib/src/domain/repository/interface/movie_interface.dart
@@ -1,0 +1,6 @@
+import '../../entity/movie_entity.dart';
+
+abstract class MovieInterface {
+  Future<Movie> getMovieById(int id);
+  Future<List<Movie>> getMovies();
+}
