--- lib/utils/strings.dart
@@ -1,15 +1,0 @@
-class Strings {
-  // Movie List
-  static const baseUrl = "https://image.tmdb.org/t/p/w500/";
-  static const jsonGenresLocation = "assets/json/genres.json";
-  static const jsonMoviesLocation = "assets/json/movies.json";
-
-  // Movie Information
-  static const releaseText = "RELEASE DATE: ";
-  static const likedText = "LIKE  -  ";
-
-  // Snapshot messages
-  static const snapshotError = "ERROR";
-  static const snapshotEmpty = "EMPTY SNAPSHOT";
-  static const snapshotOther = "ERROR NOT FOUND";
-}
